removed unnecessary api call to self every route change
